+++ xbmc/cores/VideoPlayer/Process/gbm/CMakeLists.txt
@@ -1,7 +1,4 @@
-set(SOURCES ProcessInfoGBM.cpp
-            VideoBufferDRMPRIME.cpp)
-
-set(HEADERS ProcessInfoGBM.h
-            VideoBufferDRMPRIME.h)
+set(SOURCES ProcessInfoGBM.cpp)
+set(HEADERS ProcessInfoGBM.h)
 
 core_add_library(processGBM)
